+++ flatlaf-jide-oss/src/main/java/com/formdev/flatlaf/jideoss/FlatJideOssDefaultsAddon.java
@@ -67,8 +67,12 @@ public class FlatJideOssDefaultsAddon
 
 			for( Map.Entry<Object, Object> e : defaults.entrySet() ) {
 				Object key = e.getKey();
-				if( key instanceof String && ((String)key).startsWith( "Jide" ) )
+				if( key instanceof String &&
+					(((String)key).startsWith( "Jide" ) ||
+					 ((String)key).equals( "Resizable.resizeBorder" )) )
+				{
 					jideDefaults.put( key, e.getValue() );
+				}
 			}
 		}